requirements.txt
certifi==2020.12.5
chardet==4.0.0
flatten-json==0.1.7
idna==2.10
prometheus-client==0.9.0
PyYAML==5.3.1
requests==2.25.1
requests-mock==1.8.0
six==1.15.0
urllib3==1.26.2
